Created numeric cve ids
